scrapegraph-py/pyproject.toml
[project]
name = "scrapegraph_py"
version = "1.12.2"
description = "ScrapeGraph Python SDK for API"
authors = [
    { name = "Marco Vinciguerra", email = "mvincig11@gmail.com" },
    { name = "Lorenzo Padoan", email = "lorenzo.padoan977@gmail.com" }
]


license = "MIT"
readme = "README.md"
homepage = "https://scrapegraphai.com/"
repository = "https://github.com/ScrapeGraphAI/scrapegraph-sdk"
documentation = "https://github.com/ScrapeGraphAI/scrapegraph-sdk/tree/main/scrapegraph-py#readme"
keywords = [
    "ai",
    "artificial intelligence",
    "gpt",
    "machine learning",
    "nlp",
    "natural language processing",
    "openai",
    "scraping",
    "web scraping tool",
    "webscraping",
    "graph",
    "sdk",
    "api"
    ]
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10,<4.0"

dependencies = [
    "requests>=2.32.3",
    "pydantic>=2.10.2",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.10",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
]

[project.optional-dependencies]
docs = ["sphinx==6.0", "furo==2024.5.6"]

[tool.uv]
managed = true
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-mock==3.14.0",
    "pylint>=3.2.5",
    "pytest-asyncio>=0.23.8",
    "aioresponses>=0.7.7",
    "responses>=0.25.3",
    "pytest-sugar>=1.0.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "isort>=5.13.2",
    "pre-commit>=4.0.1",
    "types-setuptools>=75.6.0.20241126",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.46",
    "mkdocstrings-python>=1.12.2",
    "poethepoet>=0.31.1",
    "twine>=6.1.0",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["F", "E", "W", "C"]
ignore = ["E203", "E501", "C901"]  # Ignore conflicts with Black and function complexity

[tool.mypy]
python_version = "3.10"
strict = true
disallow_untyped_calls = true
ignore_missing_imports = true

[build-system]
requires = ["hatchling==1.26.3"]
build-backend = "hatchling.build"

[tool.poe.tasks]
pylint-local = "pylint scrapegraph_py/**/*.py"
pylint-ci = "pylint --disable=C0114,C0115,C0116,C901 --exit-zero scrapegraph_py/**/*.py"
